feat: allow login to non-db contracts
Before this commit, each time updateAccount was triggered - i.e. a user
logged-in - the VM checked if the user connected - created an access key
- to the DB contract configured in config.contractId.

This is a strong constraint, since it does not allow users to use the VM
alongside other smart contracts with the wallet selector. Indeed, if a
user would login to any contract, and then render a Widget, the user would
be disconnected automatically.

This commit removes this constraint.